backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.4.0
httpx==0.27.0
openai==1.12.0
cohere==4.47
sentence-transformers==2.3.1
torch==2.2.0
tiktoken==0.6.0
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
markdown==3.5.2
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
nltk==3.8.1
rank-bm25==0.2.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
celery==5.3.6
redis==5.0.1
